requirements.txt
nltk==3.8.1
numpy==2.1.1
opencv_python==4.10.0.84
Pillow==10.4.0
pygame==2.6.0
scikit_learn==1.5.2
tensorflow==2.17.0
tensorflow_intel==2.17.0
